% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_all_subjects}
\alias{download_all_subjects}
\title{Download All Subjects' Data from a Project}
\usage{
download_all_subjects(project_id, format = "zip", dest_dir = getwd())
}
\arguments{
\item{project_id}{The ID of the project (e.g., "TEST").}

\item{format}{The format of the downloaded files (default is "zip").}

\item{dest_dir}{The destination directory where the files will be saved (default is current working directory).}
}
\value{
Invisibly returns a list of responses if successful. Stops execution if any download fails.
}
\description{
Downloads zip archives for all subjects within a specified project from the XNAT server.
}
\examples{
\dontrun{
  authenticate_xnat()
  download_all_subjects(
    project_id = "TEST",
    format = "zip",
    dest_dir = "~/Downloads/TEST_Subjects"
  )
}
}
