% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{download_xnat_files}
\alias{download_xnat_files}
\title{Download Files from XNAT}
\usage{
download_xnat_files(url, username, password, dest_file)
}
\arguments{
\item{url}{The full URL to the file resource you want to download.}

\item{username}{Your XNAT username.}

\item{password}{Your XNAT password.}

\item{dest_file}{The destination path where the downloaded file will be saved.}
}
\value{
Invisibly returns the response object if successful. Stops execution if the download fails.
}
\description{
Downloads files from the XNAT server using basic authentication, following redirects,
and disabling SSL verification to match the behavior of a successful `wget` command.
}
\examples{
\dontrun{
download_xnat_files(
  url = "https://rrinid.rotman-baycrest.on.ca/spred/data/projects/BuBr_M190_BA/subjects/BuBr_M190_BA_14972/experiments/BuBr_M190_BA_14972_MRI_20240108/scans/ALL/resources/DICOM/files?format=zip",
  username = "buchsbaum",
 pa
  dest_file = "BuBr_M190_BA_MRI_14972__20240108_Scan_ALL.zip"
)
}
}
