% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{list_experiment_files}
\alias{list_experiment_files}
\title{List Files in an Experiment Resource}
\usage{
list_experiment_files(project_id, subject_id, experiment_id, resource)
}
\arguments{
\item{project_id}{The project identifier.}

\item{subject_id}{The subject identifier.}

\item{experiment_id}{The experiment identifier.}

\item{resource}{The resource label.}
}
\value{
A tibble of class \code{xnat_files}.
}
\description{
Retrieves files from an experiment-level resource.
}
\examples{
\dontrun{
files <- list_experiment_files(
  project_id = "MyProject",
  subject_id = "Subject001",
  experiment_id = "Exp001",
  resource = "SNAPSHOTS"
)
}

}
