% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_select}
\alias{search_select}
\title{Add Fields to Search Builder}
\usage{
search_select(builder, ...)
}
\arguments{
\item{builder}{A search builder object.}

\item{...}{Field IDs to include in the results.}
}
\value{
The modified search builder (for chaining).
}
\description{
Add Fields to Search Builder
}
