% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_where}
\alias{search_where}
\title{Add Criterion to Search Builder}
\usage{
search_where(builder, field, comparison, value)
}
\arguments{
\item{builder}{A search builder object.}

\item{field}{Field ID to filter on.}

\item{comparison}{Comparison type.}

\item{value}{Value to compare.}
}
\value{
The modified search builder (for chaining).
}
\description{
Add Criterion to Search Builder
}
