% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{with_xnat_client}
\alias{with_xnat_client}
\title{Temporarily use a client as the global session}
\usage{
with_xnat_client(client, expr)
}
\arguments{
\item{client}{An \code{xnat_client}.}

\item{expr}{Expression to evaluate.}
}
\value{
The value of \code{expr}.
}
\description{
This helper makes existing code written for the global-session workflow work
with an explicit \code{xnat_client}.
}
