% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{xnat_client}
\alias{xnat_client}
\title{Create an \code{xnat_client}}
\usage{
xnat_client(
  base_url,
  username = NULL,
  password = NULL,
  ssl_verify = TRUE,
  jsession = NULL
)
}
\arguments{
\item{base_url}{Base URL of the XNAT server.}

\item{username}{Username.}

\item{password}{Password.}

\item{ssl_verify}{Whether to verify SSL certificates. Default \code{TRUE}.}

\item{jsession}{Optional JSESSIONID value.}
}
\value{
An \code{xnat_client}.
}
\description{
Low-level constructor for client objects. Prefer \code{xnat_connect()} for most
use cases.
}
