% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{xnat_current_client}
\alias{xnat_current_client}
\title{Get the current global-session client}
\usage{
xnat_current_client()
}
\value{
An \code{xnat_client} or \code{NULL}.
}
\description{
Returns an \code{xnat_client} populated from the package environment (set by
\code{authenticate_xnat()}), or \code{NULL} if not authenticated.
}
