% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{xnat_server}
\alias{xnat_server}
\title{Get current XNAT server URL}
\usage{
xnat_server(client = NULL)
}
\arguments{
\item{client}{Optional \code{xnat_client}. If \code{NULL}, uses the global session.}
}
\value{
The current server URL or NULL if not authenticated.
}
\description{
Get current XNAT server URL
}
\examples{
\dontrun{
xnat_server()
}

}
