% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{xnat_username}
\alias{xnat_username}
\title{Get current username}
\usage{
xnat_username()
}
\value{
The current username or NULL if not authenticated.
}
\description{
Get current username
}
\examples{
\dontrun{
xnat_username()
}

}
