% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnatR-package.R
\docType{package}
\name{xnatR-package}
\alias{xnatR}
\alias{xnatR-package}
\title{xnatR: Interface with XNAT Systems}
\description{
An R package for interacting with XNAT systems via the REST API. Provides functions to authenticate, list projects/subjects/experiments/scans, search, and download data from XNAT neuroimaging archives.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/xnatR}
  \item Report bugs at \url{https://github.com/bbuchsbaum/xnatR/issues}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
