% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate_xnat}
\alias{authenticate_xnat}
\title{Authenticate with XNAT and Cache Credentials}
\usage{
authenticate_xnat(
  base_url = NULL,
  username = NULL,
  password = NULL,
  token = NULL,
  ssl_verify = TRUE
)
}
\arguments{
\item{base_url}{The base URL of the XNAT server. If `NULL`, loaded from config.}

\item{username}{The username for XNAT authentication. If `NULL`, loaded from config.}

\item{password}{The password for XNAT authentication. Overrides config if provided.}

\item{token}{An API token for XNAT authentication. Overrides config if provided.}

\item{ssl_verify}{Logical. Whether to verify SSL certificates. Defaults to `TRUE`.}
}
\value{
Invisibly returns `TRUE` if authentication is successful. Stops execution if it fails.
}
\description{
Authenticates with the XNAT server using credentials from the config file or provided directly.
Upon successful authentication, caches the credentials for use in other functions.
}
