% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_subject}
\alias{download_subject}
\title{Download Data for a Subject}
\usage{
download_subject(project_id, subject_id, format = "zip", dest_dir = getwd())
}
\arguments{
\item{project_id}{The ID of the project.}

\item{subject_id}{The ID of the subject.}

\item{format}{The format of the downloaded files (default is "zip").}

\item{dest_dir}{The destination directory where the files will be saved (default is current working directory).}
}
\value{
Invisibly returns a list of responses if successful. Stops execution if any download fails.
}
\description{
Downloads all experiments and scans for a given subject within a specified project from the XNAT server.
}
\examples{
\dontrun{
  authenticate_xnat()
  download_subject(
    project_id = "TEST",
    subject_id = "SUBJECT_ID",
    format = "zip",
    dest_dir = "~/Downloads/TEST_Subjects/SUBJECT_ID"
  )
}
}
