% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{initialize_config}
\alias{initialize_config}
\title{Initialize Configuration File for xnatR}
\usage{
initialize_config()
}
\value{
Invisibly returns TRUE if the file is created successfully.
}
\description{
Creates a configuration file template for storing XNAT credentials.
The file is saved in the user's home directory as '.xnatR_config.yml'.
}
