% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{is_authenticated}
\alias{is_authenticated}
\title{Check if authenticated}
\usage{
is_authenticated()
}
\value{
TRUE if credentials are stored, FALSE otherwise.
}
\description{
Check if authenticated
}
\examples{
\dontrun{
if (is_authenticated()) {
  list_projects()
}
}

}
