% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_projects.R
\name{list_projects}
\alias{list_projects}
\title{List XNAT Projects}
\usage{
list_projects()
}
\value{
A data frame containing project details (e.g., ID, Label, Description).
}
\description{
Retrieves a list of projects from the authenticated XNAT server.
}
\examples{
\dontrun{
  authenticate_xnat()
  projects <- list_projects()
  print(projects)
}
}
