% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_scans.R
\name{list_scans}
\alias{list_scans}
\title{List XNAT Scans}
\usage{
list_scans(project_id, subject_id, experiment_id)
}
\arguments{
\item{project_id}{The ID of the project (e.g., "TEST").}

\item{subject_id}{The ID of the subject (e.g., "1").}

\item{experiment_id}{The ID of the experiment (e.g., "MR1").}
}
\value{
A data frame containing scan details (e.g., ID, Label, Type).
}
\description{
Retrieves a list of scans from a specific experiment, subject, and project on the authenticated XNAT server.
}
\examples{
\dontrun{
  authenticate_xnat()
  scans <- list_scans(project_id = "TEST", subject_id = "1", experiment_id = "MR1")
  print(scans)
}
}
