% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_subjects.R
\name{list_subjects}
\alias{list_subjects}
\title{List XNAT Subjects}
\usage{
list_subjects(project_id)
}
\arguments{
\item{project_id}{The ID of the project (e.g., "TEST").}
}
\value{
A data frame containing subject details (e.g., ID, Label, Gender, Age).
}
\description{
Retrieves a list of subjects from a specific project on the authenticated XNAT server.
}
\examples{
\dontrun{
  authenticate_xnat()
  subjects <- list_subjects(project_id = "TEST")
  print(subjects)
}
}
