% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_projects.R
\name{search_projects}
\alias{search_projects}
\title{Search XNAT Projects by Substring}
\usage{
search_projects(substring)
}
\arguments{
\item{substring}{A character string to search for within project IDs or Names.}
}
\value{
A data frame of projects that match the substring.
}
\description{
Retrieves projects containing a specific substring in their ID or Name.
}
\examples{
\dontrun{
  authenticate_xnat()
  results <- search_projects(substring = "Test")
  print(results)
}
}
