% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{xnat_search_builder}
\alias{xnat_search_builder}
\title{Search Builder - Fluent Interface}
\usage{
xnat_search_builder(root_type)
}
\arguments{
\item{root_type}{The XSI type to search.}
}
\value{
A search builder object.
}
\description{
Creates a search builder object for constructing complex searches
with a fluent API.
}
\examples{
\dontrun{
# Build and execute a search using fluent API
results <- xnat_search_builder("xnat:mrSessionData") |>
  search_select("xnat:mrSessionData/ID", "xnat:mrSessionData/label") |>
  search_where("xnat:mrSessionData/project", "EQUALS", "MyProject") |>
  search_where("xnat:mrSessionData/date", "GREATER_THAN", "2023-01-01") |>
  search_execute()
}

}
